requirements.txt

requests
kafka-python
pyspark
pymongo